IGNITE-14167 Simplify reconnecting, fix affinity topology change detection.
