+++ tests/config/log4j.xml.jinja2
@@ -33,7 +33,6 @@
         </RollingFile>
     </Appenders>
     <Loggers>
-        <Logger name="org.apache.ignite.internal.processors.odbc.ClientListenerNioListener" level="debug"/>
         <Root level="info">
             <AppenderRef ref="CONSOLE"/>
             <AppenderRef ref="FILE"/>
