+++ tox.ini
@@ -17,12 +17,6 @@
 skipsdist = True
 envlist = py{36,37,38}-{no-ssl,ssl,ssl-password}
 
-[travis]
-python =
-  3.6: py36-{no-ssl,ssl,ssl-password}
-  3.7: py37-{no-ssl,ssl,ssl-password}
-  3.8: py38-{no-ssl,ssl,ssl-password}
-
 [testenv]
 passenv = TEAMCITY_VERSION IGNITE_HOME
 envdir = {homedir}/.virtualenvs/pyignite-{envname}
